apps/ai-service/requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==11.2.1
requests==2.32.3
torch==2.2.2
torchvision==0.17.2
